pyproject.toml
[project]
name = "keertana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.19.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.3",
    "notebook>=7.4.3",
    "numpy>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "xlrd>=2.0.1",
]
